TECH-325 : clean code and refacto java
